[SRVCOM-1069] Stop stamping defaults into serving and eventing
