+++ knative-operator/pkg/common/util.go
@@ -3,8 +3,6 @@ package common
 import (
 	"strings"
 
-	corev1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/api/resource"
 	"k8s.io/apimachinery/pkg/runtime/schema"
 	"k8s.io/apimachinery/pkg/types"
 	"sigs.k8s.io/controller-runtime/pkg/client"
@@ -82,30 +80,6 @@ func SetAnnotations(annotations map[string]string) mf.Transformer {
 		}
 		return nil
 	}
-}
-
-func EnsureContainerMemoryLimit(s *operatorv1alpha1.CommonSpec, containerName string, memory resource.Quantity) {
-	for i, v := range s.Resources {
-		if v.Container == containerName {
-			if v.Limits == nil {
-				v.Limits = corev1.ResourceList{}
-			}
-			if _, ok := v.Limits[corev1.ResourceMemory]; ok {
-				return
-			}
-			v.Limits[corev1.ResourceMemory] = memory
-			s.Resources[i] = v
-			return
-		}
-	}
-	s.Resources = append(s.Resources, operatorv1alpha1.ResourceRequirementsOverride{
-		Container: containerName,
-		ResourceRequirements: corev1.ResourceRequirements{
-			Limits: corev1.ResourceList{
-				corev1.ResourceMemory: memory,
-			},
-		},
-	})
 }
 
 // EnqueueRequestByOwnerAnnotations is a common function to enqueue reconcile requests for resources.
